% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno_batch.R
\name{AnnoCNVBatch}
\alias{AnnoCNVBatch}
\alias{MergeAnno}
\title{The CNV batch annotation}
\usage{
AnnoCNVBatch(cns, FUN, ..., n = 2)

MergeAnno(annoList)
}
\arguments{
\item{cns}{The standard cns format.}

\item{FUN}{\code{AnnoCNVClinCore()}, \code{AnnoCNVPopuCore()}, and \code{ANnoCNVGeneCore()}.}

\item{...}{Parameters from the \code{FUN} functions}

\item{n}{The number of threads for parallel computation.}

\item{annoList}{A \code{list} containing ClinGen annotation for each called CNV. The 1st element is a \code{data.frame} or \code{NULL}. The 2nd element and the rest are \code{character} vectors.}
}
\value{
A merged \code{list}.
}
\description{
Batch annotation of CNVs
}
\details{
\code{AnnoCNVBatch()}: Annotation of multiple CNVs to ClinVar/CliGen, population, and gene databases.
}
\examples{
require('magrittr')
data(CNVdb)

cnsFile <- system.file('extdata', 'example.cnvkit', package = 'CNVanno')
cns <- cnsFile \%>\% read.cns \%>\% FilterCNS

## ClinGenGRCh38
AnnoCNVBatch(cns, AnnoCNVClinCore, CNVdb$ClinGenGRCh38)

## ClinVarGRCh38
AnnoCNVBatch(cns, AnnoCNVClinCore, CNVdb$ClinVarGRCh38,
             typeColName = 'Type',
             sigColName = 'ClinicalSignificance')
## DGV
AnnoCNVBatch(cns, AnnoCNVPopuCore, CNVdb$DGV,
             'gain_frequence', 'loss_frequence',
             'variantsubtype')

## DECIPHER_POPULATION
AnnoCNVBatch(cns, AnnoCNVPopuCore, CNVdb$DECIPHER_POPULATION)

## ExAC_POPULATION
AnnoCNVBatch(cns, AnnoCNVPopuCore, CNVdb$ExAC_POPULATION, NA, NA, 'type')

## OMIM
AnnoCNVBatch(cns, AnnoCNVGeneCore, CNVdb$OMIMGRCh38)

## ClinGen_TriHaplo
AnnoCNVBatch(cns, AnnoCNVGeneCore, CNVdb$ClinGen_TriHaploGRCh38)

## DECIPHER_Haplo
AnnoCNVBatch(cns, AnnoCNVGeneCore, CNVdb$DECIPHER_Haplo)

## ExAC_pLI
AnnoCNVBatch(cns, AnnoCNVGeneCore, CNVdb$ExAC_pLI)

}
\keyword{internal}
