% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/anno_batch.R
\docType{methods}
\name{AnnoCNVBatch}
\alias{AnnoCNVBatch}
\alias{AnnoCNVBatch,CoreCNV,function-method}
\title{The CNV batch annotation}
\usage{
AnnoCNVBatch(core, FUN, ..., n)

\S4method{AnnoCNVBatch}{CoreCNV,`function`}(core, FUN, ..., n)
}
\arguments{
\item{core}{A \code{CoreCNV} object.}

\item{FUN}{\code{AnnoCNVClinCore()}, \code{AnnoCNVPopuCore()}, or \code{ANnoCNVGeneCore()} in this package.}

\item{...}{Additional parameters passed to \code{FUN}.}

\item{n}{The number of threads for parallel computation.}
}
\description{
Batch annotation of CNVs
}
\details{
Annotation of multiple CNVs to ClinVar/ClinGen, population, and gene databases.
}
\examples{
data(CNVdb)
data(kit)

## ClinGenGRCh37
AnnoCNVBatch(kit, AnnoSVCore, CNVdb$ClinGenGRCh37,
             reciprate = 0.5, typerate = 0.7, n = 2)

## ClinVarGRCh37
AnnoCNVBatch(kit, AnnoSVCore, CNVdb$ClinVarGRCh38,
             reciprate = 0.5, typerate = 0.7, n = 2)

## DGVGRCh37
AnnoCNVBatch(kit, AnnoCNVPopuCore, CNVdb$DGVGRCh37,
             reciprate = 0.5, typerate = 0.7, n = 2)

## DECIPHER_POPULATION
AnnoCNVBatch(kit, AnnoCNVPopuCore, CNVdb$DECIPHER_POPULATION,
             reciprate = 0.5, typerate = 0.7, n = 2)

## ExAC_POPULATION
AnnoCNVBatch(kit, AnnoCNVPopuCore, CNVdb$ExAC_POPULATION,
             reciprate = 0.5, typerate = 0.7, n = 2)

## OMIM
AnnoCNVBatch(kit, AnnoCNVGeneCore, CNVdb$OMIMGRCh38, n = 2)

## ClinGen_TriHaplo
AnnoCNVBatch(kit, AnnoCNVGeneCore, CNVdb$ClinGen_TriHaploGRCh37, n = 2)

## DECIPHER_Haplo
AnnoCNVBatch(kit, AnnoCNVGeneCore, CNVdb$DECIPHER_Haplo, n = 2)

## ExAC_pLI
AnnoCNVBatch(kit, AnnoCNVGeneCore, CNVdb$ExAC_pLI, n = 2)

## RefGeneGRCh37
AnnoCNVBatch(kit, AnnoCNVGeneCore, CNVdb$RefGeneGRCh37, n = 2)
}
\keyword{internal}
