% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunanno.R
\name{AnnoCNVGeneRefGene2DDG2P}
\alias{AnnoCNVGeneRefGene2DDG2P}
\title{Annotation DECIPHER_DDG2Pdb (GRCh37) from the RefGene (GRCh37)}
\usage{
AnnoCNVGeneRefGene2DDG2P(refgeneTable, DDG2Pdb)
}
\arguments{
\item{refgeneTable}{A \code{data.frame} of annotated RefGene table.}

\item{DDG2Pdb}{The DECIPHER_DDG2Pdb database.}
}
\value{
A \code{list} of annotated DECIPHER_DDG2Pdb
}
\description{
Annotation CNV with the DECIPHER_DDG2Pdb (GRCh37) database from the RefGene (GRCh37)
}
\details{
Step1: Normal annotation CNV with the RefGene (GRCh37) database.

Step2: Use Entrez Gene ID as keys to combine RefGene and DECIPHER_DDG2Pdb database.
}
\examples{
require('magrittr')
data(CNVdb)
data(hg38cyto)

cnsFile <- system.file('extdata', 'example.cnvkit', package = 'CNVanno')
cns <- cnsFile \%>\% read.cnvkit \%>\% filter.cnvkit

cnsRefGene <- AnnoCNVBatch(cns, AnnoCNVGeneCore, CNVdb$RefGeneGRCh37)
cnsDDG2P <- AnnoCNVGeneRefGene2DDG2P(cnsRefGene[[1]], CNVdb$DECIPHER_DDG2P)
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
