% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno_utilities.R
\name{AnnoCNVOverlap_}
\alias{AnnoCNVOverlap_}
\alias{AnnoCNVType_}
\alias{AnnoGainLossRateCheck_}
\title{Annotation utilities}
\usage{
AnnoCNVOverlap_(cnsSingle, annoSingle, mutualRate = 0.5)

AnnoCNVType_(cnsSingle, annoSingle, typeColName, typeRate = 0.7)

AnnoGainLossRateCheck_(detectType, annoType)
}
\arguments{
\item{cnsSingle}{A \code{character} vector indicating a single CNV.}

\item{annoSingle}{A \code{data.frame} indicating a single database.}

\item{mutualRate}{A \code{numeric} value indicating the mutual overlap rate.}

\item{typeColName}{A \code{character} string indicating the gain/loss column name.}

\item{typeRate}{A \code{numeric} value indicating the gain/loss rate.}

\item{detectType}{A \code{character} string, either "gain" or "loss".}

\item{annoType}{A \code{character} vector containing "gain" and "loss".}
}
\value{
\itemize{
  \item \code{AnnoCNVOverlap_()}: A \code{data.frame} with selected rows. The row number may be zero.
  \item \code{AnnoCNVType_()}: Same as the \code{AnnoCNVOverlap_()} function.
 \item \code{AnnoGainLossRateCheck_()}: A \code{logic} vector.
}
}
\description{
Internal annotation functions
}
\details{
\itemize{
  \item \code{AnnoCNVOverlap_()}: Annotation of single CNV with one database with mutual overlap rate.
  \item \code{AnnoCNVType_()}: Annotation of single CNV with one database with gain and loss percentage. The input database should be the output of \code{AnnoCNVOverlap_()} function.
  \item \code{AnnoGainLossRateCheck_()}: Check the gain/loss annotation.
}
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
\keyword{internal}
