% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{bl_cytoband}
\alias{bl_cytoband}
\title{Blacklist of cytoband}
\usage{
bl_cytoband(cyto, extend = 500000L)
}
\arguments{
\item{cyto}{A \code{tbl_df} represents the cytoband.  A code{tbl_df} with at least three columns. 1st column is chromosome in the same format with \code{core}. 2nd and 3rd columns are start and end positions. 4th column is the cytoband. Other columns can be included but will not be used.}

\item{extend}{A \code{integer} value indicates the extended length in both direction.}
}
\value{
A \code{tbl_df} represents the black list of cytoband.
}
\description{
Filter the cytoband.
}
\details{
Construct the blacklist of cytoband. The \code{extend} can be set as 100,000 500,000 or 1000,000. The "acen", "gvar", and "stalk" will be kept in the black list.
}
\examples{
data(hg19cyto)

bl_cytoband(hg19cyto, extend = 5e5L)
}
\references{
\href{http://penncnv.openbioinformatics.org/en/latest/user-guide/annotation/#filtering-cnv-calls-by-user-specified-criteria}{cytoband extend}
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
