% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno_clin.R
\name{AnnoCNVClinCore}
\alias{AnnoCNVClinCore}
\alias{AnnoBenignPathoCheck_}
\title{The ClinGen and ClinVar database annotation}
\usage{
AnnoCNVClinCore(corerow, annodb, reciprate = 0.5, typerate = 0.7)

AnnoBenignPathoCheck_(annoSig)
}
\arguments{
\item{corerow}{A row of the CNV in a \code{tbl_df} form.}

\item{annodb}{A \code{tbl_df} indicating a single annotation database. In \code{AnnoCNVClinCore()}, it should at least contains "chromosome", "start", "end", "type" (gain or loss), and "clinical_significance" columns.}

\item{reciprate}{The reciprocal overlap rate between two CNVs.}

\item{typerate}{A \code{numeric} value indicating the gain/loss rate.}

\item{annoSig}{A \code{character} vector containing "benign" and "pathogentic".}
}
\value{
\itemize{
   \item \code{AnnoCNVClinCore()}: A \code{list}
}

A \code{list} summary the clinical significance.
}
\description{
CNV ClinGen and ClinVar database annotation
}
\details{
\itemize{
  \item \code{AnnoCNVClinCore()}: Annotation of single CNV to the ClinGen and ClinVar database.
}
}
\examples{
data(kit)

## ClinGen
AnnoCNVClinCore(kit@coreCNV[3, ], CNVdb$ClinGenGRCh37)

## ClinVar
AnnoCNVClinCore(kit@coreCNV[3, ], CNVdb$ClinVarGRCh37)
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
\keyword{internal}
