% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/anno_combine.R
\docType{methods}
\name{CombinePara}
\alias{CombinePara}
\alias{CombinePara,CoreCNV,RawCNV,numeric-method}
\title{Combine parameters from \code{RawCNV}}
\usage{
CombinePara(core, raw, reciprate, n, ...)

\S4method{CombinePara}{CoreCNV,RawCNV,numeric}(core, raw, reciprate, n,
  ...)
}
\arguments{
\item{core}{A \code{CoreCNV} object.}

\item{raw}{A \code{RawCNV} object.}

\item{reciprate}{The reciprocal overlap rate between two CNVs.}

\item{n}{The number of threads for parallel computation.}

\item{...}{Additional parameters.}
}
\value{
A \code{CoreCNV} object.
}
\description{
Combine parameters from the raw CNV outputs
}
\details{
Merge the parameters from the raw CNV outpus.
}
\examples{
require('magrittr')
require('dplyr')
data(nator)
data(kit)

mergecnv <- Merge(list(nator, kit), reciprate = 0.5, n = 2)

kitraw <- system.file('extdata', 'example.cnvkit', package = 'CNVanno') \%>\% read_cnvkit
kitraw@params \%<>\% select(log2 : depth)

natorraw <- system.file('extdata', 'example.cnvnator', package = 'CNVanno') \%>\% read_cnvnator
natorraw@params \%<>\% select(normalized_RD, q0) \%>\% rename(nordepth = normalized_RD)

mergecnv \%>\% CombinePara(kitraw, 1e-05, n = 2) \%>\% CombinePara(natorraw, 1e-05, n = 2)

}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
\keyword{internal}
