% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno_combine.R
\name{compressPara_}
\alias{compressPara_}
\alias{combineRow_}
\title{Internal functions for CombinePara}
\usage{
compressPara_(raw)

combineRow_(corerow, rawpara, reciprate)
}
\arguments{
\item{raw}{A \code{RawCNV}.}

\item{corerow}{A row of the CNV in a \code{tbl_df} form.}

\item{rawpara}{A \code{tbl_df}}

\item{reciprate}{The reciprocal overlap rate between two CNVs.}
}
\value{
\itemize{
  \item \code{compressPara_()}: A \code{tbl_df}.
  \item \code{combineRow_()}: A \code{string}.
}
}
\description{
CombinePara internal functions
}
\details{
\itemize{
  \item \code{compressPara_()}: Compress the parameters of \code{RawCNV} into one column and combine it with the `rawCNV` slot.
  \item \code{combineRow_()}: Combine each row of the \code{RawCNV}.
}
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
\keyword{internal}
