% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CoreCNV-class}
\alias{CoreCNV-class}
\title{This class represents the CNV structure used for further analysis.}
\description{
This class represents the CNV structure used for further analysis.
}
\section{Slots}{

\describe{
\item{\code{coreCNV}}{An \code{tbl_df} from the \code{tibble} package. It contains five columns: 1st is the chromosome, 2nd is the CNV start position, 3rd is the CNV end position, 4th is the CNV type ("gain", "loss", "normal"), 5th is the CNV calling method.}
}}

\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
