% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno_template.R
\name{CrossRegionGeneTable}
\alias{CrossRegionGeneTable}
\title{Combine CNVregion and CNVgene tables.}
\usage{
CrossRegionGeneTable(regionTable, geneTable)
}
\arguments{
\item{regionTable}{A \code{tbl_df} of processed region table, which often comes from the \code{SunCNVregion()} in this package.}

\item{geneTable}{A \code{tbl_df} of gene table, which often comes from the \code{SunCNVgene()} in this package.}
}
\value{
A list containing two \code{tbl_df} represents the processed \code{regionTable} and \code{geneTable} respectively.
}
\description{
Combine CNVregion and CNVgene tables according to Sun's format.
}
\details{
\itemize{
  \item \code{CrossRegionGeneTable()}: Merge region and gene tables.
}
}
\examples{
require('magrittr')
data(CNVdb)
data(mergecnv)
data(hg19cyto)

refGene <- AnnoCNVBatch(mergecnv, AnnoCNVGeneCore, CNVdb$RefGeneGRCh37, n = 2)
gdbList <- list()
gdbList[[1]] <- AnnoCNVGeneRefGene2OMIM(refGene[[1]], CNVdb$OMIMGRCh38)
gdbList[[2]] <- AnnoCNVBatch(mergecnv, AnnoCNVGeneCore, CNVdb$ClinGen_TriHaploGRCh37, n = 2)[[1]]
gdbList[[3]] <- AnnoCNVBatch(mergecnv, AnnoCNVGeneCore, CNVdb$DECIPHER_Haplo, n = 2)[[1]]
gdbList[[4]] <- AnnoCNVBatch(mergecnv, AnnoCNVGeneCore, CNVdb$ExAC_pLI, n = 2)[[1]]
gdbList[[5]] <- AnnoCNVGeneRefGene2DDG2P(refGene[[1]], CNVdb$DECIPHER_DDG2P)

geneTable <- SunCNVgene(gdbList)
regionTable <- SunCNVregion(mergecnv, hg19cyto, n = 2)
grList <- CrossRegionGeneTable(regionTable, geneTable)
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
