% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvfilter.R
\name{filter.cnvkit}
\alias{filter.cnvkit}
\title{CNS filter}
\usage{
filter.cnvkit(cns, cngain = 2, cnloss = 2, log2value = 0.5,
  depth = 0.01, sexchrom = TRUE)
}
\arguments{
\item{cns}{The standard \code{CNVkit} output.}

\item{cngain}{A \code{numeric} value as the gain copy number threshold, and the default value is 2.}

\item{cnloss}{A \code{numeric} value as the loss copy number threshold, and the default value is 2.}

\item{log2value}{A \code{numeric} value as the absolute log2value threshold, and the default value is absolute 0.5.}

\item{depth}{A \code{numeric} value as the sequence depth, and the default value is 0.01}

\item{sexchrom}{A \code{logic} value whether filter sex chromosomes (X and Y), the default value is \code{TRUE}.}
}
\value{
A filtered cns format
}
\description{
Filter the CNV file generated from \code{CNVkit} and \code{CNVnator}
}
\details{
For the \code{CNVkit}
\itemize{
  \item 1. filter the cns with cn number is 2 (no variance).
  \item 2. filter log2 value, and the default threshold is absolute 0.5.
  \item 3. filter low sequence depth.
  \item 4. filter X/Y chromosomes
}

For the \code{CNVnator}
\itemize{
  \item 1. 
}
}
\examples{
require('magrittr')

cnsFile <- system.file('extdata', 'example.cnvkit', package = 'CNVanno')
cnsFiltered <- cnsFile \%>\% read.cnvkit \%>\% filter.cnvkit

}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
