% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_cnvkit}
\alias{filter_cnvkit}
\title{CNV filter}
\usage{
filter_cnvkit(RawCNVkit, cngain = 2, cnloss = 2, dep = 0.01,
  sexchrom = TRUE)
}
\arguments{
\item{RawCNVkit}{The standard \code{CNVkit} output as an \code{RawCNV} object.}

\item{cngain}{A \code{numeric} value as the gain copy number threshold, and the default value is 2.}

\item{cnloss}{A \code{numeric} value as the loss copy number threshold, and the default value is 2.}

\item{dep}{A \code{numeric} value as the sequence depth, and the default value is 0.01}

\item{sexchrom}{A \code{logic} value whether filter sex chromosomes (X and Y), the default value is \code{TRUE}.}
}
\value{
A filtered \code{RawCNV} object.
}
\description{
Filter the CNV file generated from \code{CNVkit}
}
\details{
For the \code{CNVkit}
\itemize{
  \item 1. filter the \code{cns} with cn number.
  \item 2. filter low sequence depth.
  \item 3. filter X/Y chromosomes
}
}
\examples{
require('magrittr')

kitf <- system.file('extdata', 'example.cnvkit', package = 'CNVanno') \%>\% read_cnvkit \%>\% filter_cnvkit

}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
