% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/filter.R
\docType{methods}
\name{FilterBlacklist}
\alias{FilterBlacklist}
\alias{FilterBlacklist,CoreCNV,tbl_df,numeric,integer,integer-method}
\title{Filter \code{CoreCNV} according to a give blacklist}
\usage{
FilterBlacklist(core, blacklist, overlaprate, shortlen, gap, n, ...)

\S4method{FilterBlacklist}{CoreCNV,tbl_df,numeric,integer,integer}(core,
  blacklist, overlaprate, shortlen, gap, n, ...)
}
\arguments{
\item{core}{a \code{CoreCNV} object.}

\item{blacklist}{A code{tbl_df} with at least three columns. 1st column is chromosome in the same format with \code{core}. 2nd and 3rd columns are start and end positions. Other columns can be included but will not be used. The blacklist must be reduced (use \code{ReduceRegionChr()}) and sorted (use \code{SortRegionChr()}).}

\item{overlaprate}{The threshold of overlap rate (overlaplen/CNVlen). The CNVs with smaller than or equal to this rate will be filtered.}

\item{shortlen}{A code{integer}. The CNVs with length shorter than or equal to this value will be filtered.}

\item{gap}{An \code{integer} indicate the allowed length between two regions (at same chromosome with same type).}

\item{n}{The number of threads for parallel computation.}

\item{...}{Additional parameters.}
}
\value{
A \code{CoreCNV} object.
}
\description{
Filtering \code{CoreCNV}.
}
\details{
Filer the \code{CoreCNV} according black lists. Multiple black lists can be applied as the example show below. q
}
\examples{
library('magrittr')
data(hg19cyto)
data(hg19bl)

nator <- system.file('extdata', 'example.cnvnator', package = 'CNVanno') \%>\%
  read_cnvnator \%>\%
  filter_cnvnator \%>\%
  Segment(gap = 10L)

## filter based on cytoband blacklist
hg19cytobl <- bl_cytoband(hg19cyto)
natorf <- FilterBlacklist(nator, hg19cytobl, overlaprate = 0.5, shortlen = 10L, gap = 0L, n = 2)

## more filter based on pre-built blacklist
natorf <- FilterBlacklist(natorf, hg19bl, overlaprate = 0.5, shortlen = 10L, gap = 1000L, n = 2)
}
\references{
\href{http://penncnv.openbioinformatics.org/en/latest/user-guide/annotation/#filtering-cnv-calls-by-user-specified-criteria}{cytoband extend}
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
\keyword{internal}
