% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filterSeg_}
\alias{filterSeg_}
\alias{filterRow_}
\title{Internal functions for filtering}
\usage{
filterSeg_(regionf, rateMat, chr, type)

filterRow_(corerow, blacklist, overlaprate)
}
\arguments{
\item{regionf}{A \code{numeric} vector with length two, and the value of 1st position is smaller/equal to that of 2nd position.}

\item{rateMat}{The output of \code{OverlapRegionRate()} in this package.}

\item{chr}{A \code{string} indicating the chromosome, like "chr1", "chr2".}

\item{type}{A code{string} "gain", "loss", or "normal"}

\item{corerow}{A row of the CNV in a \code{tbl_df} form.}
}
\value{
A \code{tbl_df} of filtered fragments.
}
\description{
Filtering internal functions.
}
\details{
\itemize{
  \item \code{filterSeg_()}: For the blacklist (<= `overlaprate`) in the same chromosome, segment the regions of input `regionf`. If the `regionf` wrap up the `bl` (====~~~~===), `regionf` is cut into two segments. The minimum left and right regions are kept.
  \item \code{filterRow_()}: Filter input one cnv (one row).
}
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
\keyword{internal}
