% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno.R
\name{FindCyto}
\alias{FindCyto}
\title{Find cytoband}
\usage{
FindCyto(cnvcore, cyto, n = 2)
}
\arguments{
\item{cnvcore}{A \code{data.frame} represent the CNV core data structure. 1st column is the chromosome like \code{chr1}, \code{chrX}, and \code{chrM}. 2nd and 3rd columns are start and end positions, respectively.}

\item{cyto}{A standard cytoband file.}

\item{n}{The number of threads for parallel computation.}
}
\value{
A \code{character} vector.
}
\description{
Find cytoband of given cns.
}
\examples{
require('magrittr')
data(hg19cyto)

cnsFile <- system.file('extdata', 'example.cnvkit', package = 'CNVanno')
cnsFilteredCyto <- cnsFile \%>\% read.cnvkit \%>\% filter.cnvkit \%>\% FindCyto(hg19cyto)

}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
