% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDef.R
\docType{data}
\name{hg19bl}
\alias{hg19bl}
\title{Blacklist of hg19 genome.}
\format{A \code{tbl_df}}
\description{
Combination of low mapbility regions and blacklist from the 10X genome.
}
\references{
\href{ftp://hgdownload.soe.ucsc.edu/goldenPath/hg19/encodeDCC/wgEncodeMapability/wgEncodeDacMapabilityConsensusExcludable.bed.gz}{mapbility_concensus}

\href{ftp://hgdownload.soe.ucsc.edu/goldenPath/hg19/encodeDCC/wgEncodeMapability/wgEncodeDukeMapabilityRegionsExcludable.bed.gz}{mapbility_region}

\href{http://cf.10xgenomics.com/supp/genome/hg19/sv_blacklist.bed}{10X genome blacklist}
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
