% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDef.R
\docType{data}
\name{hg19cyto}
\alias{hg19cyto}
\title{Cytobands of hg19 genome from the UCSC goldenpath}
\format{A \code{tbl_df}}
\description{
The default hg19 cytobands.
}
\references{
\href{http://hgdownload.cse.ucsc.edu/goldenpath/hg19/database/}{UCSC goldenpath}
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
