% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDef.R
\docType{data}
\name{hg38bl}
\alias{hg38bl}
\title{Blacklist of hg38 genome.}
\format{A \code{tbl_df}}
\description{
Combination of low mapbility regions and blacklist from the 10X genome.
}
\references{
\href{http://cf.10xgenomics.com/supp/genome/hg38/sv_blacklist.bed}{10X genome blacklist}
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
