% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDef.R
\docType{data}
\name{hg38cyto}
\alias{hg38cyto}
\title{Cytobands of hg38 genome from the UCSC goldenpath}
\format{A \code{data.frame}}
\description{
The default hg38 cytobands.
}
\references{
\href{http://hgdownload.cse.ucsc.edu/goldenpath/hg38/database/}{UCSC goldenpath}
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
