% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDef.R
\docType{data}
\name{kit}
\alias{kit}
\title{CNVkit filtered example file}
\format{A \code{coreCNV} object}
\description{
## CNVanno version 0.2.0
kit <- system.file('extdata', 'example.cnvkit', package = 'CNVanno') %>%
  read_cnvkit %>%
  filter_cnvkit %>%
  Segment(gap = 10L)%>%
  FilterBlacklist(bl_cytoband(hg19cyto), overlaprate = 0.5, shortlen = 1000L, gap = 0L, n = 2) %>%
  FilterBlacklist(hg19bl, overlaprate = 0.5, shortlen = 1000L, gap = 1000000L, n = 2)
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
