% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/merge.R
\docType{methods}
\name{Merge}
\alias{Merge}
\alias{Merge,list,numeric-method}
\title{Merge \code{CoreCNV}}
\usage{
Merge(corelist, reciprate, n, ...)

\S4method{Merge}{list,numeric}(corelist, reciprate, n, ...)
}
\arguments{
\item{corelist}{A list of \code{CoreCNV} objects used to merge.}

\item{reciprate}{The reciprocal overlap rate between two CNVs.}

\item{n}{The number of threads for parallel computation.}

\item{...}{Additional parameters.}
}
\value{
A \code{CoreCNV} object.
}
\description{
Merge two \code{CoreCNV}.
}
\details{
Merge CNVs with reciprocal overlap regions.
}
\examples{
require('magrittr')
data(nator)
data(kit)

Merge(list(nator, kit), reciprate = 0.5, n = 2)
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
\keyword{internal}
