% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno_batch.R
\name{mergeAnno_}
\alias{mergeAnno_}
\title{Merge CNV annotation}
\usage{
mergeAnno_(annoList)
}
\arguments{
\item{annoList}{A \code{list} containing CNV annotation for each called CNV. The 1st element is a \code{data.frame} or \code{NULL}. The 2nd element and the rest are \code{character} vectors.}
}
\value{
A merged \code{list}.
}
\description{
Internal merge annotation functions
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
\keyword{internal}
