% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDef.R
\docType{data}
\name{mergecnv}
\alias{mergecnv}
\title{Merged CNVkit and CNVnator with parameters annotation}
\format{A \code{coreCNV} object}
\description{
mergecnv <- Merge(list(nator, kit), reciprate = 0.5, n = 2)
kitraw <- system.file('extdata', 'example.cnvkit', package = 'CNVanno') %>% read_cnvkit
kitraw@params %<>% select(log2 : depth)
natorraw <- system.file('extdata', 'example.cnvnator', package = 'CNVanno') %>% read_cnvnator
natorraw@params %<>% select(normalized_RD, q0) %>% rename(nordepth = normalized_RD)
mergecnv %<>% CombinePara(kitraw, 1e-05, n = 2) %>% CombinePara(natorraw, 1e-05, n = 2)
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
