% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno_template.R
\name{SunMergeGenedb}
\alias{SunMergeGenedb}
\title{Merge genes database}
\usage{
SunMergeGenedb(gdbList)
}
\arguments{
\item{gdbList}{A list containing CNV genes database. 1st element is the OMIM database, 2nd is the ClinGenTriHaplo, 3rd is the DECIPHERHaplo, 4th is the ExACpLI, 5th is the DECPHER_DDG2P}
}
\value{
A \code{data.frame}
}
\description{
The CNV summary genes table.
}
\details{
## column names for OMIM
\itemize{
  \item CNV
  \item overlap_relation
  \item gene_symbol
  \item Phenotypes
  \item Phenotypes CN
  \item MIMNumber
  \item MIMType
}
## column names for ClinGenTriHaplo
\itemize{
  \item Haploinsufficiency Score
  \item Triplosensitivity Score
}

## column names for DECIPHERHaplo
\itemize{
  \item haploinsufficiency
}

## column names for pLI
\itemize{
  \item pLI
}

\itemize{
  \item disease.name
  \item DDD.category
  \item allelic.requirement
  \item mutation.consequence
  \item phenotypes
}
}
\examples{
require('magrittr')
data(CNVdb)
data(kit)

refGene <- AnnoCNVBatch(kit, AnnoCNVGeneCore, CNVdb$RefGeneGRCh37, n = 2)
gdbList <- list()
gdbList[[1]] <- AnnoCNVGeneRefGene2OMIM(refGene[[1]], CNVdb$OMIMGRCh38)
gdbList[[2]] <- AnnoCNVBatch(kit, AnnoCNVGeneCore, CNVdb$ClinGen_TriHaploGRCh37, n = 2)[[1]]
gdbList[[3]] <- AnnoCNVBatch(kit, AnnoCNVGeneCore, CNVdb$DECIPHER_Haplo, n = 2)[[1]]
gdbList[[4]] <- AnnoCNVBatch(kit, AnnoCNVGeneCore, CNVdb$ExAC_pLI, n = 2)[[1]]
gdbList[[5]] <- AnnoCNVGeneRefGene2DDG2P(refGene[[1]], CNVdb$DECIPHER_DDG2P)

geneTable <- SunMergeGenedb(gdbList)

## cnsTable <- cns \%>\% SunCNVTable(hg38cyto, sampleType = 'proband', n = CORENUM)
## CrossRegionGeneTable(cnsTable, geneTable)
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
