% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno_template.R
\name{mergeTwoGenedb_}
\alias{mergeTwoGenedb_}
\title{Internal Sun annotation functions}
\usage{
mergeTwoGenedb_(gdb1, gdb2)
}
\arguments{
\item{gdb1, }{gdb2: Annotated gene databases.}
}
\value{
\itemize{
  \item \code{mergeTwoGenedb_()}: A \code{tbl_df}.
}
}
\description{
Internal Sun's template annotation functions.
}
\details{
\itemize{
  \item \code{mergeTwoGenedb_()}: Merge two gene results.
}
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
\keyword{internal}
