% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{mergePrepare_}
\alias{mergePrepare_}
\alias{mergeHasOverlap_}
\alias{mergeReduce_}
\alias{mergeSearch_}
\title{Internal functions for merge}
\usage{
mergePrepare_(corelist)

mergeHasOverlap_(regf, regt, reciprate, n)

mergeReduce_(reglist)

mergeSearch_(regt, reciprate, n)
}
\arguments{
\item{corelist}{A list of \code{CoreCNV} objects used to merge.}

\item{regf, }{regt A \code{tbl_df} including "start", "end", and "method" columns.}

\item{reciprate}{The reciprocal overlap rate between two CNVs.}

\item{n}{The number of threads for parallel computation.}

\item{reglist}{A \code{list} contains \code{tbl_df}.}
}
\value{
\itemize{
  \item \code{mergePrepare_()}: A \code{tbl_df} merged CNV.
  \item \code{mergeHasOverlap_()}: A \code{logical vector}.
  \item \code{mergeReduce_()}: A code{tbl_df} with one row.
}
A \code{tbl_df} of input CNVs.
}
\description{
Merge internal functions.
}
\details{
\itemize{
  \item \code{mergePrepare_()}: Prepare the input \code{CoreCNV} objects.
  \item \code{mergeHasOverlap_()}: Has passed overlap regions.
  \item \code{mergeReduce_()}: Reduce selected regions.
}
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
\keyword{internal}
