% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno_population.R
\name{AnnoCNVPopuCore}
\alias{AnnoCNVPopuCore}
\alias{AnnoPopuPercentage_}
\title{The CNV population database annotation}
\usage{
AnnoCNVPopuCore(corerow, annodb, reciprate = 0.5, typerate = 0.7)

AnnoPopuPercentage_(annoPopuPer)
}
\arguments{
\item{corerow}{A row of the CNV in a \code{tbl_df} form.}

\item{annodb}{A \code{tbl_df} indicating a single annotation database. In \code{AnnoCNVClinCore()}, it should at least contains "chromosome", "start", "end", "type" (gain or loss), and "clinical_significance" columns. In \code{AnnoCNVPopuCore()}, it should at least contains "chromosome", "start", "end", "type" (gain or loss), may contain "gain_frequency"and "loss_frequency".}

\item{reciprate}{The reciprocal overlap rate between two CNVs.}

\item{typerate}{A \code{numeric} value indicating the gain/loss rate.}

\item{annoPopuPer}{A \code{numeric} vector containing the percentage of CNV in the population.}
}
\value{
\itemize{
  \item \code{AnnoCNVPopuCore()}: A \code{list}
}

A \code{character} vector summary the CNV percentage.
}
\description{
Annotation core function for the DECIPHER_POPULATION, DGV, and ExAC_POPULATION databases
}
\details{
\itemize{
  \item \code{AnnoCNVPopuCore()}: Annotation of single CNV to one population database.
}
}
\examples{
data(kit)

## DECIPHER_POPULATION
AnnoCNVPopuCore(kit@coreCNV[4, ], CNVdb$DECIPHER_POPULATION)

## DGV
AnnoCNVPopuCore(kit@coreCNV[4, ], CNVdb$DGVGRCh37)

## ExAC_POPULATION
AnnoCNVPopuCore(kit@coreCNV[3, ], CNVdb$ExAC_POPULATION)

}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
\keyword{internal}
