% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw.R
\name{read.cns}
\alias{read.cns}
\alias{write.cns}
\title{Standard read in cns file}
\usage{
read.cns(cnspath)

write.cns(cns, cnspath)
}
\arguments{
\item{cnspath}{The path of cns file.}

\item{cns}{The standard cns format.}
}
\value{
\itemize{
  \item \code{read.cns()}: A \code{data.frame}.
  \item \code{write.cns()}: A table separated \code{txt} format file.
}
}
\description{
Read and write cns file
}
\details{
\itemize{
  \item \code{read.cns()}: Read in cns file generated from \code{CNVkit} as a \code{data.frame}.
  \item \code{write.cns()}: Write standard cns format file as table.
}
}
\examples{
require('magrittr')

cnsFile <- system.file('extdata', 'example.cnvkit', package = 'CNVanno')
cnsMat <- cnsFile \%>\% read.cns
\dontrun{
## write cns file
write.cns(cnsMat, 'cnsMat.txt')
}

}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
