% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw.R
\name{read.cnvkit}
\alias{read.cnvkit}
\alias{write.cnv}
\title{Standard read and write multiple cnv files}
\usage{
read.cnvkit(cnvpath)

write.cnv(cnv, cnspath)
}
\arguments{
\item{cnvpath}{The path of cnv file.}

\item{cns}{The standard cns format.}
}
\value{
\itemize{
  \item \code{read.cnvkit()}: A \code{data.frame}.
  \item \code{write.cnvkit()}: A table separated \code{txt} format file.
}
}
\description{
Read and write cnvkit cns file
}
\details{
\itemize{
  \item \code{read.cnvkit()}: Read in cns file generated from \code{CNVkit} as a \code{data.frame}.
  \item \code{write.cnvkit()}: Write standard cnv format file as table.
}
}
\examples{
require('magrittr')

cnsFile <- system.file('extdata', 'example.cnvkit', package = 'CNVanno')
cnsMat <- cnsFile \%>\% read.cnvkit
\dontrun{
## write cns file
write.cnvkit(cnsMat, 'cnsMat.txt')
}

}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
