% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segmentcore_}
\alias{segmentcore_}
\title{Internal functions for segment}
\usage{
segmentcore_(corecnv, gap)
}
\arguments{
\item{corecnv}{A \code{tbl_df} contains "chromosome", "start", "end", "type", and "method" columns.}

\item{gap}{An \code{integer} indicate the allowed length between two regions (at same chromosome with same type).}
}
\value{
A \code{tbl_df}.
}
\description{
Segment internal functions
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
\keyword{internal}
