% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segMergeType_}
\alias{segMergeType_}
\alias{segMergeChr_}
\alias{segMerge_}
\title{Internal functions for segmentation}
\usage{
segMergeType_(cnv, gap, chr, type)

segMergeChr_(cnv, gap, chr)

segMerge_(cnv, gap)
}
\arguments{
\item{cnv}{A \code{tbl_df} from the \code{rawCNV} slot of a \code{RawCNV} object.}

\item{gap}{An \code{integer} indicate the allowed length between two regions (at same chromosome with same type).}

\item{chr}{A \code{string} indicating the chromosome, like "chr1", "chr2".}

\item{type}{A code{string} "gain", "loss", or "normal"}
}
\value{
A \code{tbl_df}
}
\description{
Segmentation internal functions.
}
\details{
\itemize{
  \item \code{segMergeType_()}: Merege CNVs in the same chromosome in the same type.
  \item \code{segMergeChr_()}: Merege CNVs in the same chromosome.
  \item \code{segMerge_()}: Merge CNVs.
}
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
\keyword{internal}
