% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno_template.R
\name{SunCNVregion}
\alias{SunCNVregion}
\title{Sun's summary CNV table}
\usage{
SunCNVregion(core, cyto, sampleType = "proband", ...)
}
\arguments{
\item{core}{A \code{CoreCNV} object.}

\item{cyto}{A \code{tbl_df} represents the cytoband.  A code{tbl_df} with at least three columns. 1st column is chromosome in the same format with \code{core}. 2nd and 3rd columns are start and end positions. 4th column is the cytoband. Other columns can be included but will not be used.}

\item{sampleType}{The sample type, like "proband", "mother", "father".}

\item{...}{Parameters passed to the \code{Cytoband()} in this package.}
}
\value{
A \code{tbl_df} contains Sun's CNV region format.
}
\description{
The CNV region summary table according to Sun's format.
}
\details{
\itemize{
  \item CNV
  \item size(100kb)
  \item type
  \item method
  \item cytoband
  \item sample
}
}
\examples{
require('magrittr')
data(hg19cyto)
data(mergecnv)

SunCNVregion(mergecnv, hg19cyto, n = 2)
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
