% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunanno.R
\name{SunCNVTable}
\alias{SunCNVTable}
\title{Sun's summary CNV table}
\usage{
SunCNVTable(cns, cyto, sampleType = "proband", n = 2)
}
\arguments{
\item{sampleType}{The sample type, like "proband", "mother", "father".}
}
\value{
A \code{data.frame} of Sun's cns format.
}
\description{
The CNV summary table from Sun.
}
\details{
\itemize{
  \item REGION
  \item SIZE
  \item RATIO
  \item COPY
  \item TYPE
  \item CYTOBAND
  \item SAMPLE
  \item DERIVED
}
}
\examples{
require('magrittr')
data(hg38cyto)

cnsFile <- system.file('extdata', 'example.cnvkit', package = 'CNVanno')
cnsSummarySun <- cnsFile \%>\% read.cnvkit \%>\% filter.cnvkit \%>\% SunCNVTable(hg38cyto)
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
