% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno_utilities.R
\name{AnnoCNVOverlap_}
\alias{AnnoCNVOverlap_}
\alias{AnnoCNVType_}
\alias{FormatCorerow_}
\title{Annotation utilities}
\usage{
AnnoCNVOverlap_(corerow, annodb, reciprate = 0.5)

AnnoCNVType_(corerow, annodb, typerate = 0.7)

FormatCorerow_(corerow)
}
\arguments{
\item{corerow}{A row of the CNV in a \code{tbl_df} form.}

\item{annodb}{A \code{tbl_df} indicating a single annotation database. In \code{AnnoSVCore()}, it should at least contains "chromosome", "start", "end", "type" (gain or loss), and "clinical_significance" columns. In \code{AnnoCNVPopuCore()}, it should at least contains "chromosome", "start", "end", "type" (gain or loss), may contain "gain_frequency"and "loss_frequency". In \code{AnnoCNVGeneCore()}, it should at least contains "chromosome", "start" and "end" columns.}

\item{reciprate}{The reciprocal overlap rate between two CNVs.}

\item{typerate}{A \code{numeric} value indicating the gain/loss rate.}
}
\value{
\itemize{
  \item \code{AnnoCNVOverlap_()}: A \code{data.frame} with selected rows. The row number may be zero.
  \item \code{AnnoCNVType_()}: Same as the \code{AnnoCNVOverlap_()} function.
 \item \code{FormatCorerow_()}: A \code{string}.
}
}
\description{
Internal annotation functions
}
\details{
\itemize{
  \item \code{AnnoCNVOverlap_()}: Annotation of single CNV with one database with mutual overlap rate.
  \item \code{AnnoCNVType_()}: Annotation of single CNV with one database with gain and loss percentage. The input database should be the output of \code{AnnoCNVOverlap_()} function.
  \item \code{FormatCorerow_()}: Format single CNV.
}
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
\keyword{internal}
