% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno_sv.R
\name{AnnoSVCore}
\alias{AnnoSVCore}
\alias{AnnoBenignPathoCheck_}
\title{The dbVar, ClinGen, and ClinVar databases annotation}
\usage{
AnnoSVCore(corerow, annodb, reciprate = 0.5, typerate = 0.7)

AnnoBenignPathoCheck_(annoSig)
}
\arguments{
\item{corerow}{A row of the CNV in a \code{tbl_df} form.}

\item{annodb}{A \code{tbl_df} indicating a single annotation database. In \code{AnnoSVCore()}, it should at least contains "chromosome", "start", "end", "type" (gain or loss), and "clinical_significance" columns. In \code{AnnoCNVPopuCore()}, it should at least contains "chromosome", "start", "end", "type" (gain or loss), may contain "gain_frequency"and "loss_frequency". In \code{AnnoCNVGeneCore()}, it should at least contains "chromosome", "start" and "end" columns.}

\item{reciprate}{The reciprocal overlap rate between two CNVs.}

\item{typerate}{A \code{numeric} value indicating the gain/loss rate.}

\item{annoSig}{A \code{character} vector containing "benign" and "pathogentic".}
}
\value{
\itemize{
   \item \code{AnnoSVCore()}: A \code{list}
}

A \code{list} summary the clinical significance.
}
\description{
CNV dbVar, ClinGen, and ClinVar databases annotation
}
\details{
\itemize{
  \item \code{AnnoSVCore()}: Annotation of single CNV to the dbVar,  ClinGen, and ClinVar databases.
}
}
\examples{
data(CNVdb)
data(kit)

## dbVar
AnnoSVCore(kit@coreCNV[3, ], CNVdb$dbVarGRCh37)

## ClinGen
AnnoSVCore(kit@coreCNV[3, ], CNVdb$ClinGenGRCh37)

## ClinVar
AnnoSVCore(kit@coreCNV[3, ], CNVdb$ClinVarGRCh37)
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
\keyword{internal}
