% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno_clin.R
\name{AnnoCNVClinCore}
\alias{AnnoCNVClinCore}
\alias{AnnoBenignPathoCheck_}
\title{The ClinGen and ClinVar database annotation}
\usage{
AnnoCNVClinCore(cnsSingle, annoClin, typeColName = "variant_call_type",
  sigColName = "clinical_significance", mutualRate = 0.5,
  typeRate = 0.7)

AnnoBenignPathoCheck_(annoSig)
}
\arguments{
\item{cnsSingle}{A \code{character} vector indicating a single CNV.}

\item{annoClin}{A \code{data.frame} of the ClinGen and ClinVar database.}

\item{typeColName}{A \code{character} string indicating the gain/loss column name.}

\item{sigColName}{A \code{character} string indicating the benign and pathogentic column name.}

\item{mutualRate}{A \code{numeric} value indicating the mutual overlap rate.}

\item{typeRate}{A \code{numeric} value indicating the gain/loss rate.}

\item{annoSig}{A \code{character} vector containing "benign" and "pathogentic".}
}
\value{
\itemize{
   \item \code{AnnoCNVClinCore()}: A \code{list}
}

A \code{list} summary the clinical significance.
}
\description{
CNV ClinGen and ClinVar database annotation
}
\details{
\itemize{
  \item \code{AnnoCNVClinCore()}: Annotation of single CNV to the ClinGen and ClinVar database.
}
}
\examples{
require('magrittr')
data(CNVdb)

cnsFile <- system.file('extdata', 'example.cnvkit', package = 'CNVanno')
cns <- cnsFile \%>\% read.cnvkit \%>\% FilterCNS

## ClinGen
AnnoCNVClinCore(cns[1, ], CNVdb$ClinGenGRCh38)

## ClinVar
AnnoCNVClinCore(cns[1, ],
                CNVdb$ClinVarGRCh38,
                typeColName = 'Type',
                sigColName = 'ClinicalSignificance')
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
\keyword{internal}
