% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDef.R
\docType{data}
\name{CNVdb}
\alias{CNVdb}
\title{Annotation databases of CNV}
\format{A \code{list}}
\description{
\itemize{
  \item ClinGenGRCh38 2017-04-01
  \item ClinGenGRCh37 2017-04-01
  \item ClinGenNCBI36 2017-04-01
  \item ClinVarGRCh38 2018-08-21
  \item ClinVarGRCh37 2018-08-21
  \item ClinVarNCBI36 2018-08-21
  \item ClinGen_TriHaploGRCh38 2018-08-24
  \item ClinGen_TriHaploGRCh37 2018-08-24
  \item DECIPHER_Haplo
  \item DECIPHER_POPULATION
  \item DECIPHER_DDG2P 2018-07-14
  \item DGVGRCh38 2016-08-31
  \item DGVGRCh37 2016-05-15
  \item ExAC_POPULATION
  \item ExAC_pLI
  \item OMIMGRCh38 2018-07-13
  \item RefGeneGRCh37
}
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
