% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno_gene.R
\name{AnnoCNVGeneCore}
\alias{AnnoCNVGeneCore}
\alias{AnnoGeneOverlap_}
\alias{OverlapRelation_}
\title{The CNV gene database annotation}
\usage{
AnnoCNVGeneCore(cnsSingle, annoGene)

AnnoGeneOverlap_(cnsSingle, annoSingle)

OverlapRelation_(rateMat)
}
\arguments{
\item{cnsSingle}{A \code{character} vector indicating a single CNV.}

\item{annoGene}{A \code{data.frame} of the OMIM, ClinGen_HaploTriplo, ExAC_plI, and DECIPHER_Haplo databases.}

\item{rateMat}{A \code{numeric} matrix with two columns. 1st column is the rate of overlapped region on overlap region, and 2nd column is the rate of annotated gene/region.}
}
\value{
A \code{list}

A \code{data.frame} summary annotated genes/regions.

A \code{character vector} summary coverage of genes/regions.
}
\description{
Annotation core function for the OMIM, ClinGen_TriHaplo, ExAC_plI, and DECIPHER_Haplo
}
\details{
Annotation of single CNV to one gene database
}
\examples{
require('magrittr')
data(CNVdb)

cnsFile <- system.file('extdata', 'example.cnvkit', package = 'CNVanno')
cns <- cnsFile \%>\% read.cns \%>\% FilterCNS

##OMIM
AnnoCNVGeneCore(cns[1, ], CNVdb$OMIMGRCh38)

}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
\keyword{internal}
