% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/filter.R
\docType{methods}
\name{FilterBlacklist}
\alias{FilterBlacklist}
\alias{FilterBlacklist,CoreCNV,tbl_df,numeric-method}
\title{Filter \code{CoreCNV} according to a give blacklist}
\usage{
FilterBlacklist(core, blacklist, overlaprate, n, ...)

\S4method{FilterBlacklist}{CoreCNV,tbl_df,numeric}(core, blacklist,
  overlaprate, n, ...)
}
\arguments{
\item{core}{a \code{CoreCNV} object.}

\item{blacklist}{A code{tbl_df} with at least three columns. 1st column is chromosome in the same format with \code{core}. 2nd and 3rd columns are start and end positions. Other columns can be included but will not be used. The blacklist must be reduced (use \code{ReduceRegionChr()}) and sorted (use \code{SortRegionChr()}).}

\item{overlaprate}{The threshold of overlap rate (overlaplen/CNVlen).}

\item{n}{The number of threads for parallel computation.}

\item{...}{Additional parameters.}
}
\value{
A \code{CoreCNV} object.
}
\description{
Filtering \code{CoreCNV}.
}
\details{
Filer the \code{CoreCNV} according black lists.
}
\examples{
library('magrittr')
data(hg19cyto)

nator <- system.file('extdata', 'example.cnvnator', package = 'CNVanno') \%>\%
  read_cnvnator \%>\%
  filter_cnvnator \%>\%
  Segment(gap = 10L)

## natorf <- FilterBlacklist(nator, bl_cytoband(hg19cyto), overlaprate = 0.5, n = 1)
}
\references{
\href{http://penncnv.openbioinformatics.org/en/latest/user-guide/annotation/#filtering-cnv-calls-by-user-specified-criteria}{cytoband extend}
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
\keyword{internal}
