% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno.R
\name{FindCyto}
\alias{FindCyto}
\title{Find cytoband}
\usage{
FindCyto(cns, cyto, n = 2)
}
\arguments{
\item{cyto}{A standard cytoband file.}

\item{n}{The number of threads for parallel computation.}
}
\value{
A \code{character} vector.
}
\description{
Find cytoband of given cns.
}
\examples{
require('magrittr')
data(hg19cyto)

cnsFile <- system.file('extdata', 'example.cnvkit', package = 'CNVanno')
cnsFilteredCyto <- cnsFile \%>\% read.cnvkit \%>\% FilterCNS \%>\% FindCyto(hg38cyto)

}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
