% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunanno.R
\name{SunMergeGenesdb}
\alias{SunMergeGenesdb}
\alias{MatchString}
\alias{GenedbPost}
\alias{CrossRegionGeneTable}
\title{Merge genes database}
\usage{
SunMergeGenesdb(gdbList)

MatchString(inputStr, searchStr)

GenedbPost(mergeTable)

CrossRegionGeneTable(regionTable, geneTable)
}
\arguments{
\item{gdbList}{A list containing CNV genes database. 1st element is the OMIM database, 2nd is the ClinGenTriHaplo, 3rd is the DECIPHERHaplo, 4th is the ExACpLI.}

\item{inputStr}{A \code{string}.}

\item{searchStr}{A \code{string vector}.}

\item{mergeTable}{A \code{data.frame} of processed gene table.}

\item{regionTable}{A \code{data.frame} of processed region table, which often comes from the \code{SunCNVTable} functions.}

\item{geneTable}{A \code{data.frame} of gene table, which often comes from the \code{SunMergeGenesdb} functions.}
}
\value{
A \code{data.frame}

A processed \code{data.frame}.

\code{NULL} or the index

A list containing two \code{data.frame} represents the processed \code{regionTable} and \code{geneTable} respectively.
}
\description{
The CNV summary genes table for Sun.
}
\details{
## column names for OMIM
\itemize{
  \item CNV
  \item OverlapRelation
  \item Gene.Symbols
  \item Phenotypes
  \item Phenotypes.CN
  \item genetic???
  \item Mim.Number
}
## column names for ClinGenTriHaplo
\itemize{
  \item Haploinsufficiency.Score
  \item Haploinsufficiency.PMID1
  \item Haploinsufficiency.PMID2
  \item Haploinsufficiency.PMID3
  \item Triplosensitivity.Score
  \item Triplosensitivity.PMID1
  \item Triplosensitivity.PMID2
  \item Triplosensitivity.PMID3
  \item Date.Last.Evaluated
  \item Loss.phenotype.OMIM.ID
  \item Triplosensitive.phenotype.OMIM.ID
}

## column names for DECIPHERHaplo
\itemize{
  \item haploinsufficiency
}

## column names for pLI
\itemize{
  \item pLI
}

\itemize{
  \item disease.name
  \item DDD.category
  \item allelic.requirement
  \item mutation.consequence
  \item phenotypes
  \item organ.specificity.list
}

## example column (has blanks)
## chr3:162819884-162903642	CNVinGene	HDLCQ5	[High density lipoprotein cholesterol level QTL 5], 610761 (2)			610761	40				40				2014-03-12			0.020830666	0.003227266288348
}
\examples{
require('magrittr')
data(CNVdb)
data(hg38cyto)

cnsFile <- system.file('extdata', 'example.cnvkit', package = 'CNVanno')
cns <- cnsFile \%>\% read.cns \%>\% FilterCNS

gdbList <- list()
gdbList[[1]] <- AnnoCNVBatch(cns, AnnoCNVGeneCore, CNVdb$OMIMGRCh38)[[1]]
gdbList[[2]] <- AnnoCNVBatch(cns, AnnoCNVGeneCore, CNVdb$ClinGen_TriHaploGRCh38)[[1]]
gdbList[[3]] <- AnnoCNVBatch(cns, AnnoCNVGeneCore, CNVdb$DECIPHER_Haplo)[[1]]
gdbList[[4]] <- AnnoCNVBatch(cns, AnnoCNVGeneCore, CNVdb$ExAC_pLI)[[1]]
cnsRefGene <- AnnoCNVBatch(cns, AnnoCNVGeneCore, CNVdb$RefGeneGRCh37)
gdbList[[5]] <- AnnoCNVGeneRefGene2DDG2P(cnsRefGene[[1]], CNVdb$DECIPHER_DDG2P)

geneTable <- SunMergeGenesdb(gdbList)

cnsTable <- cns \%>\% SunCNVTable(hg38cyto, sampleType = 'proband', n = CORENUM)
CrossRegionGeneTable(cnsTable, geneTable)
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}

Yulong Niu \email{yulong.niu@hotmail.com}

Yulong Niu \email{yulong.niu@hotmail.com}

Yulong Niu \email{yulong.niu@hotmail.com}
}
\keyword{internal}
