% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{OverlapRegionRate}
\alias{OverlapRegionRate}
\alias{OverlapRegion}
\alias{SortRegion}
\alias{SortRegionChr}
\alias{ReduceRegion}
\alias{ReduceRegionChr}
\title{Overlapped regions}
\usage{
OverlapRegionRate(regionf, regionMat)

OverlapRegion(regionf, regionMat, extend = 100L)

SortRegion(regionMat)

SortRegionChr(regionMatChr)

ReduceRegion(regionMat, gap)

ReduceRegionChr(regionMatChr, gap)
}
\arguments{
\item{regionf}{A \code{numeric} vector with length two, and the value of 1st position is smaller/equal to that of 2nd position.}

\item{regionMat}{A \code{tbl_df} with at least 2 columns named "start" and "end". In the \code{OverlapRegionRate()} and \code{ReduceRegion()} function, use the \code{SortRegion()} to sort the \code{regionMat} at first.}

\item{extend}{A \code{integer} value indicates the extended length in both direction.}

\item{regionMatChr}{A \code{tbl_df} with at least columns named "chromosome", "start", and "end".}

\item{gap}{An \code{integer} indicate the allowed length between two regions (at same chromosome with same type).}
}
\value{
\itemize{
  \item \code{OverlapRegionRate()}: A \code{logic} value.
  \item \code{OverlapRegion()}: A \code{numeric} matrix with four columns. 1st column is the overlap rate of `regionf`, 2nd column is the overlap rate of `regionMat`, 3rd and 4th columns are intersection start and end regions.
  \item \code{SortRegion()}: The same object as the input.
  \item \code{SortRegionChr()}: The same object as the input.
  \item \code{ReduceRegion()}: The same object as the input.
  \item \code{ReduceRegion()}: The same object as the input but only "chromosome", "start", and "end" columns are kept.
}
}
\description{
Find regions have intersections
}
\details{
\itemize{
  \item \code{OverlapRegionRate()}: Check if extended regions have intersections with overlap rates.
  \item \code{OverlapRegion()}: Check regions have interactions
  \item \code{SortRegion()}: Sort the "start" and "end" columns resulting in the "start" column is less than or equal to that of the "end" column. At last, the "start" column is also sorted.
  \item \code{SortRegionChr()}: Sort the input which is split by the "chromosome" column.
  \item \code{ReduceRegion()}: Merge concatenated regions.
  \item \code{ReduceRegionChr()}: Merge concatenated regions by the "chromosome" column.
}
}
\examples{
require('magrittr')

tMat <- tibble(start = c(1L, 103L, 111L, 49L, 10L),
               end = c(101L, 112L, 1000L, 8L, 86L)) \%>\%
  SortRegion
tReg <- c(100L, 110L)

## overlapped regions
OverlapRegion(tReg, tMat, extend = 0L)

## overlap rate
OverlapRegionRate(tReg, tMat)

## reduce region
rMat <- tibble(start = c(1, 8, 14, 15, 19, 34, 40),
               end = c(12, 13, 19, 29, 24, 35, 46)) \%>\%
  SortRegion
ReduceRegion(rMat, gap = 0L)
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
