% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{OverlapRegionRate}
\alias{OverlapRegionRate}
\alias{OverlapRegion}
\alias{SortMat}
\title{Overlapped regions}
\usage{
OverlapRegionRate(regionf, regionMat)

OverlapRegion(regionf, regionMat, extend = 100)

SortMat(regionMat)
}
\arguments{
\item{regionf}{A \code{numeric} vector with length two, and the value of 1st position is smaller/equal to that of 2nd position.}

\item{regionMat}{A \code{numeric} matrix with 2 columns. In the \code{OverlapRegionRate()} function, the rows should have overlap with the input \code{reginf}. The 1st column should be equal or smaller than the 2nd column, otherwise use the \code{SortMat()} sort the \code{regionMat} at first.}

\item{extend}{A \code{numeric} value indicates the extended length in both direction.}
}
\value{
\itemize{
  \item \code{OverlapRegionRate()}: A \code{logic} value.
  \item \code{OverlapRegion()}: A \code{numeric} matrix with two columns. 1st column is the overlap rate of `regionf`, 2nd column is the overlap rate of `regionMat`.
  \item \code{SortMat()}: A same object as the input \code{regionMat}
}
}
\description{
Find regions have intersections
}
\details{
\itemize{
  \item \code{OverlapRegionRate()}: Check if extended regions have intersections with overlap rates.
  \item \code{OverlapRegion()}: Check regions have interactions
  \item \code{SortMat()}: Sort the 1st and 2nd columns resulting in 1st column is less than or equal to that of 2nd column.
}
}
\examples{
require('magrittr')

tMat <- matrix(c(1, 101, 103, 112, 111, 1000, 8, 149, 10, 86), ncol = 2, byrow = TRUE)
tReg <- c(100, 110)

## overlapped regions
inLogic <- OverlapRegion(tReg, tMat, extend = 0)

## overlap rate
OverlapRegionRate(tReg, tMat[inLogic, ])
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
