% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{OverlapRegionRate}
\alias{OverlapRegionRate}
\alias{OverlapRegion}
\alias{SortRegion}
\title{Overlapped regions}
\usage{
OverlapRegionRate(regionf, regionMat)

OverlapRegion(regionf, regionMat, extend = 100)

SortRegion(regionMat)
}
\arguments{
\item{regionf}{A \code{numeric} vector with length two, and the value of 1st position is smaller/equal to that of 2nd position.}

\item{regionMat}{A \code{tbl_df} with 2 columns. In the \code{OverlapRegionRate()} function. The 1st column should be equal or smaller than the 2nd column, otherwise use the \code{SortRegion()} sort the \code{regionMat} at first.}

\item{extend}{A \code{numeric} value indicates the extended length in both direction.}
}
\value{
\itemize{
  \item \code{OverlapRegionRate()}: A \code{logic} value.
  \item \code{OverlapRegion()}: A \code{numeric} matrix with four columns. 1st column is the overlap rate of `regionf`, 2nd column is the overlap rate of `regionMat`, 3rd and 4th columns are intersection start and end regions.
  \item \code{SortRegion()}: A same object as the input \code{regionMat}
}
}
\description{
Find regions have intersections
}
\details{
\itemize{
  \item \code{OverlapRegionRate()}: Check if extended regions have intersections with overlap rates.
  \item \code{OverlapRegion()}: Check regions have interactions
  \item \code{SortRegion()}: Sort the 1st and 2nd columns resulting in 1st column is less than or equal to that of 2nd column.
}
}
\examples{
require('magrittr')

tMat <- tibble(start = c(1, 103, 111, 49, 10), end = c(101, 112, 1000, 8, 86)) \%>\% SortRegion
tReg <- c(100, 110)

## overlapped regions
OverlapRegion(tReg, tMat, extend = 0)

## overlap rate
OverlapRegionRate(tReg, tMat)

## sort matrix at first

}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
