% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno_population.R
\name{AnnoCNVPopuCore}
\alias{AnnoCNVPopuCore}
\alias{AnnoPopuPercentage_}
\title{The CNV population database annotation}
\usage{
AnnoCNVPopuCore(cnsSingle, annoPopu,
  gainPerColName = "duplication_frequency",
  lossPerColName = "deletion_frequency", typeColName = "cnvtype",
  mutualRate = 0.5, typeRate = 0.7)

AnnoPopuPercentage_(annoPopuPer)
}
\arguments{
\item{cnsSingle}{A \code{character} vector indicating a single CNV.}

\item{gainPerColName}{A \code{character} string indicating the gain column name.}

\item{lossPerColName}{A \code{character} string indicating the loss column name.}

\item{typeColName}{A \code{character} string indicating the gain/loss column name.}

\item{mutualRate}{A \code{numeric} value indicating the mutual overlap rate.}

\item{typeRate}{A \code{numeric} value indicating the gain/loss rate.}

\item{annoPopuPer}{A \code{numeric} vector containing the percentage of CNV in the population.}

\item{annoPopo}{A \code{data.frame} of the DECIPHER_POPULATION, DGV, and ExAC_POPULATION databases.}
}
\value{
\itemize{
  \item \code{AnnoCNVPopuCore()}: A \code{list}
}

A \code{character} vector summary the CNV percentage.
}
\description{
Annotation core function for the DECIPHER_POPULATION, DGV, and ExAC_POPULATION databases
}
\details{
\itemize{
  \item \code{AnnoCNVPopuCore()}: Annotation of single CNV to one population database.
}
}
\examples{
require('magrittr')
data(CNVdb)

cnsFile <- system.file('extdata', 'example.cnvkit', package = 'CNVanno')
cns <- cnsFile \%>\% read.cnvkit \%>\% FilterCNS

## DECIPHER_POPULATION
AnnoCNVPopuCore(cns[1, ], CNVdb$DECIPHER_POPULATION)

## DGV
AnnoCNVPopuCore(cns[1, ], CNVdb$DGV, 'gain_frequence', 'loss_frequence', 'variantsubtype')

}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
\keyword{internal}
