% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw.R
\name{read_cnvkit}
\alias{read_cnvkit}
\alias{read_cnvnator}
\title{Standard read in CNV files}
\usage{
read_cnvkit(cnvpath)

read_cnvnator(cnvpath)
}
\arguments{
\item{cnvpath}{The path of cnv file.}
}
\value{
\itemize{
  \item \code{read_cnvkit()}/\code{read_cnvkit}: A \code{RawCNV}.
}
}
\description{
Read in CNV file
}
\details{
\itemize{
  \item \code{read_cnvkit()}: Read in CNV generated from \code{CNVkit} (version 0.9.3). The \code{*.cns} file should be preprocessed by the \code{call} command in the \code{CNVkit}.
  \item \code{read_cnvnator()}: Read in CNV generated from \code{CNVnator} (version 0.3.3).
}
}
\examples{
require('magrittr')

kit <- system.file('extdata', 'example.cnvkit', package = 'CNVanno') \%>\% read_cnvkit

nator <- system.file('extdata', 'example.cnvnator', package = 'CNVanno') \%>\% read_cnvnator

\dontrun{
## write CNV file
write.cnvkit(cnsMat, 'cnsMat.txt')
}

}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
