% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/segment.R
\docType{methods}
\name{Segment}
\alias{Segment}
\alias{Segment,RawCNV,integer-method}
\title{Segmentation}
\usage{
Segment(raw, interlen, ...)

\S4method{Segment}{RawCNV,integer}(raw, interlen = 100L, ...)
}
\arguments{
\item{raw}{a \code{RawCNV} object.}

\item{interlen}{An code{integer number} indicate the integer length between CNV (at same chromosome with same type).}

\item{...}{Additional parameters.}
}
\value{
A \code{numeric matrix} object.
}
\description{
Segmentation of \code{RawCNV}.
}
\details{
Merge the concatenated CNV at the same chromosome with the same type (gain and loss).
}
\examples{
library('magrittr')

kitf <- system.file('extdata', 'exampleseg.cnvkit', package = 'CNVanno') \%>\% read_cnvkit \%>\% filter_cnvkit
kitfseg <- Segment(kitf, interlen = 10L)
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
\keyword{internal}
